dargahpardkht/requirement.txt
annotated-types==0.7.0
asgiref==3.8.1
contextlib2==21.6.0
Django==5.2.2
django-ninja==1.4.3
django-ninja-extra==0.30.1
fire==0.7.0
injector==0.22.0
markdown-it-py==3.0.0
mdurl==0.1.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
PyYAML==6.0.2
rav==0.0.9
rich==14.0.0
sqlparse==0.5.3
termcolor==3.1.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
